+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 ### Removed/Deprecated
 
 ### Added
+- Bars can now be built by `beginTime` instead of `endTime`
 
 
 ## 0.18 (released May 15, 2025)